完成classreading和扫描加载的联系。通过解析class不用jvm加载class文件就能够辨别出哪个类有@Component注解。 节省了大量类的加载时间和空间

+++ src/main/java/classreader/basicinfo/attribute/AttributeInfo.java
@@ -25,7 +25,7 @@ public class AttributeInfo extends BasicInfo {
     }
 
     public void read(InputStream in) {
-        length = (int) U4.read(in);
+        length = U4.read(in);
         info = new short[(int) length];
         for (int i = 0; i < length; i++) {
             info[i] = U1.read(in);
